+++ CHANGELOG.md
@@ -3,7 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### ✅ Added
+- Added factory method for customizing the message list container's modifier
 
 # [4.33.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.33.0)
 _June 09, 2023_